Certificate registry
Write a simple contract in Solidity that represents a registry of certificates. The contract must check whether the user that tries to add a certificate is the contract’s owner. If this is so the certificate is added, otherwise – no. The other functionality should allow a user to check whether a certain certificate is in the registry. In this exercise the certificates will be hashed and the hashed value will be passed to the appropriate function.