Accounts struct
Write a simple contract in Solidity that keeps track of users’ names, addresses and e-mails. The contract must check whether the user that tries to get users’ information is the contract’s owner – other users must not be able to consult the records. The contract should provide functionality that allows to anyone to add their address’ information.