Simple token
Write a contract in Solidity that represents a simple token. The initial supply must be set at contract’s creation and this amount must be allocated to the address that creates the contract. Add a functionality that allows for transfers of tokens between the address of the contract’s creator and other addresses